Several new scripts plus extra beta version scripts

+++ beta/buildnewversiongit.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 # Build a new MonoDevelop version from the git repos
-export VERSION=4.3.4
-export PREVVERSION=4.3.2
-export PACKAGEVERSION=4.0
-export PREVPACKAGEVERSION=4.0
+export VERSION=5.1.0.462
+export PREVVERSION=5.0.1.3
+export PACKAGEVERSION=5
+export PREVPACKAGEVERSION=5
+export TAG_VERSION=${VERSION}
+export GITTAG=monodevelop-${TAG_VERSION}
 # REPO specifies the launchpad project where the package should end up. This should probably
 # be monodevelop or monodevelop-beta
 [ -z "${REPO}" ] && REPO=monodevelop-beta
 
-[ -z "${MODULES}" ] && MODULES="monodevelop monodevelop-debugger-gdb monodevelop-database"
+[ -z "${MODULES}" ] && MODULES="monodevelop monodevelop-database"
 WHERE=$(pwd)
 
 # Prevent suffix adding to package names
@@ -29,7 +31,7 @@ echo "**************** Updating MonoDevelop repo *******************"
 git reset --hard
 git clean -dxf -e debian
 git fetch origin
-git checkout monodevelop-${VERSION}
+git checkout ${GITTAG}
 git submodule update --init
 cd $WHERE
 
@@ -42,9 +44,9 @@ if [ ! "${NOBUILD}" ]; then
 		"monodevelop")
 			SUBDIR=main
 			;;
-		"monodevelop-debugger-gdb")
-			SUBDIR=extras/MonoDevelop.Debugger.Gdb
-			;;
+		#"monodevelop-debugger-gdb")
+		#	SUBDIR=extras/MonoDevelop.Debugger.Gdb
+		#	;;
 		"monodevelop-database")
 			SUBDIR=extras/MonoDevelop.Database
 			;;
@@ -91,12 +93,16 @@ if [ ! "${NOBUILD}" ]; then
 			SUBDIR=extras/ValaBinding
 			;;
 		esac
-		mkdir -p $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
-		cp -a $WHERE/monodevelop/$SUBDIR/* $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
-		cp -a $WHERE/monodevelop/version.config $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
-		cp -a $WHERE/monodevelop/scripts $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
 		cd $WHERE
-		tar cfj ${MODULE}-${PACKAGEVERSION}_${VERSION}.orig.tar.bz2 ${MODULE}-${PACKAGEVERSION}-${VERSION}
+		if [ -f ${MODULE}-${PACKAGEVERSION}_${VERSION}.orig.tar.bz2 ] ; then
+			tar xfj ${MODULE}-${PACKAGEVERSION}_${VERSION}.orig.tar.bz2
+		else
+			mkdir -p $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
+			cp -a $WHERE/monodevelop/$SUBDIR/* $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
+			cp -a $WHERE/monodevelop/version.config $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
+			cp -a $WHERE/monodevelop/scripts $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
+			tar cfj ${MODULE}-${PACKAGEVERSION}_${VERSION}.orig.tar.bz2 --exclude=debian ${MODULE}-${PACKAGEVERSION}-${VERSION}
+		fi
 		if [ ! -d "${MODULE}-${PACKAGEVERSION}-${VERSION}/debian" ]; then
 			cp -r ${MODULE}-${PREVPACKAGEVERSION}-${PREVVERSION}/debian ${MODULE}-${PACKAGEVERSION}-${VERSION}/
 			cd ${MODULE}-${PACKAGEVERSION}-${VERSION}/debian
@@ -109,23 +115,25 @@ if [ ! "${NOBUILD}" ]; then
 				echo
 				echo "Please adjust the patches, then exit the subshell"
 				bash
+				echo "Building ${MODULE}..."
 			fi
 			sed 's/monodevelop-'${PACKAGEVERSION}' (>= [0-9.]*)/monodevelop-'${PACKAGEVERSION}' (>= '${VERSION}')/' < /tmp/control > control
 		fi
 		# Update version information
 		if [ "$MODULE" = "monodevelop" ]; then
+			echo "Update version information"
 			cd $WHERE/monodevelop/main
 			mkdir -p build/bin
 			../scripts/configure.sh gen-buildinfo "./build/bin"
 			cd $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
-			quilt push -a
+			quilt push -a || true
 			if [ "$(quilt top)" != "buildinfo.patch" ]; then
 				quilt new buildinfo.patch
 				quilt add build/bin/buildinfo
 			fi
 			cp -a $WHERE/monodevelop/main/build/bin build/
 			quilt refresh
-			quilt pop -a
+			quilt pop -a > /dev/null
 		fi
 
 		cd $WHERE/${MODULE}-${PACKAGEVERSION}-${VERSION}
@@ -142,7 +150,9 @@ if [ ! "${NOUPLOAD}" ]; then
 	echo "******************** Uploading changes **********************"
 	for MODULE in $MODULES
 	do
-		debsign ${MODULE}-${PACKAGEVERSION}_${VERSION}-*source.changes
-		dput ppa:ermshiperete/${REPO} ${MODULE}-${PACKAGEVERSION}_${VERSION}-*source.changes
+		for f in ${MODULE}-${PACKAGEVERSION}_${VERSION}-*source.changes; do
+			debsign --no-re-sign $f
+			dput ppa:ermshiperete/${REPO} $f
+		done
 	done
 fi
